[WPE][GTK] Move `WebKitWebsiteData/memory-pressure` API test to a separate binary again https://bugs.webkit.org/show_bug.cgi?id=260256
Reviewed by Carlos Garcia Campos.

This test verifies that the network process will be killed if the memory
limit is reached.

It's done by calling `webkit_network_session_set_memory_pressure_settings()`.
The settings passed as an argument are later applied during the creation
of a new network process.

Before 259433@main a new network process was created for every network
session but now only one network process is created.

Since this test is not the first one in the test suite, the network
process was already created and the memory pressure settings have no effect.

Also, the name of the function is misleading now,
because the memory pressure settings are applied to all sessions.

The previous PR 266946@main was reverted because it caused a build failure.
It happened due to a missing header. This PR fixes it.

* Tools/TestWebKitAPI/Tests/WebKitGLib/TestNetworkProcessMemoryPressure.cpp: Added.
(serverCallback):
(MemoryPressureTest::setup):
(MemoryPressureTest::teardown):
(MemoryPressureTest::loadFailedCallback):
(MemoryPressureTest::waitUntilLoadFailed):
(testMemoryPressureSettings):
(beforeAll):
(afterAll):
* Tools/TestWebKitAPI/Tests/WebKitGLib/TestWebsiteData.cpp:
(beforeAll):
(MemoryPressureTest::setup): Deleted.
(MemoryPressureTest::teardown): Deleted.
(MemoryPressureTest::loadFailedCallback): Deleted.
(MemoryPressureTest::waitUntilLoadFailed): Deleted.
(testMemoryPressureSettings): Deleted.
* Tools/TestWebKitAPI/glib/CMakeLists.txt:

Canonical link: https://commits.webkit.org/267000@main
